termano test fe purchase

+++ src/controllers/cart.controllers.js
@@ -2,13 +2,22 @@ const catchError = require('../utils/catchError');
 const Cart = require('../models/Cart');
 const User = require('../models/User')
 const Product = require('../models/Product')
+const Category = require('../models/Category')
 
 const getAll = catchError(async(req, res) => {
-   console.log(req.user.id)
-    const results = await Cart.findAll({where:{userId:req.user.id},include:[User, Product]});
-   //{where:{userId:req.user.id}, include:[User, Product]}
-    
-    if(!results) return res.json({"messege":"empty cart"})
+ 
+    const results = await Cart.findAll({where:{userId:req.user.id},
+        include:[{
+        model:Product,
+        attributes:{exclude:['updatedAt','createdAt']},
+
+        include:[{
+            model:Category,
+            attributes:['name','id']
+        }]
+    }]
+});
+      if(!results) return res.json({"messege":"empty cart"})
     return res.json(results);
 });
 
@@ -22,14 +31,23 @@ const create = catchError(async(req, res) => {
 
 const getOne = catchError(async(req, res) => {
     const { id } = req.params;
-    const result = await Cart.findByPk(id,{include:[User, Product]});
+    const result = await Cart.findByPk(id,{
+        include:[{
+            model:Product,
+            attributes:{exclude:['updatedAt','createdAt']},
+            include:[{
+                model:Category,
+                attributes:['name','id']
+            }]
+
+        }]});
     if(!result) return res.sendStatus(404);
     return res.json(result);
 });
 
 const remove = catchError(async(req, res) => {
     const { id } = req.params;
-    const result = await Cart.destroy({ where: {id} });
+    const result = await Cart.destroy({ where: {id,userId:req.user.id} });
     if(!result) return res.sendStatus(404);
     return res.sendStatus(204).json({"message":"Delete Cart"});
 });
@@ -41,7 +59,7 @@ const update = catchError(async(req, res) => {
      delete req.body.userId
      delete req.body.productId
    }
-    const result = await Cart.update(req.body,{ where: {id}, returning: true }
+    const result = await Cart.update(req.body,{ where: {id,userId:req.user.id}, returning: true }
     );
     if(result[0] === 0) return res.sendStatus(404);
     return res.json(result[1][0]);
